Add dynamic command error detection
Depending on initial configuration, some dynamic commands are not allowed.
For example it is not allowed to set a number of B-frame higher than the
initial configuration because it would required to allocate more buffers
while encoding.

Note : The Dynamic feature APIs have slightly changed and now each function
returns a boolean value to check the validity.

Signed-off-by: Pascal EYMERY <pascal.eymery@allegrodvt.com>