Fix slicelat avc concealment case
An example of this concealment case:
[0] slice received first_mb_in_slice = 0, poc = 4
[1] slice received first_mb_in_slice = 100, poc = 4
[2] slice received first_mb_in_slice = 200, poc = 8
Here we need to conceal the previous frame by setting [1] last slice flag
Then we add a first slice to the frame with poc 8 and we are good to go.

The error was in the code setting the last slice flag of [1]. We were sending
wrong slices.
As we change the last slice flag, we go in a special case where the slice is
its own previous slice and we add special code for this part.

Signed-off-by: Kevin Grandemange <kevin.grandemange@allegrodvt.com>
Signed-off-by: Jeegar Patel <jeegar.patel@xilinx.com>