Add first draft of the sei encoder api.
Missing: check on the length of the generated sei to ensure it can be written inside
the stream buffer.

The user is for now responsible for _all_ of the sei_payload() syntax, including
the byte alignement bits, but we take care of the sei_rbsp and sei_message bits.

for the sei prefix, we search for the last configuration section and we add a
section after it with the sei payload data.

for the sei suffix, we just add a section at the end of the current section array.

(see lib_common/unittests/metadata.cpp for tests using these mechanism)

This means the api should be called only after the stream has been encoded.

This adds an options to AL_Encoder (--dummy-sei) to use the new external sei api

Signed-off-by: Kevin Grandemange <kevin.grandemange@allegrodvt.com>
Signed-off-by: Jeegar Patel <jeegar.patel@xilinx.com>